got generic puzzle generation working, and added move functions. up and down tested, left and right untested.

+++ src/utilities/structs.go
@@ -13,6 +13,7 @@ type ThreePuzzle struct {
 }
 
 type Puzzle struct {
-	size uint8
-	board []uint8
+	size int
+	board []int
 }
+
